Create navigation adapter
Created NavigationAdapter class to handle
navigation drawer item selection events.
Removed navigation item selection listener
from main activity.
